project/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
cffi==1.16.0
cryptography==41.0.5
defusedxml==0.7.1
dj-rest-auth==2.2.4
Django==4.2.3
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
dj-rest-auth==2.2.4
django-allauth==0.45.0
django-cors-headers==4.2.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
frozenlist==1.4.0
gTTS==2.3.2
multidict==6.0.4
openai==0.27.8
tqdm==4.65.0
yarl==1.9.2
python-dotenv==0.19.1
djangorestframework-simplejwt 
django-environ==0.10.0
